debug storia mombellica

+++ laboratorioMombelli/Java/java-inteldj/storia/src/main/java/com/mombelli/storia/repos/EventRepo.java
@@ -3,5 +3,5 @@ package com.mombelli.storia.repos;
 import com.mombelli.storia.entities.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepo extends JpaRepository<Event, Integer> {
+public interface EventRepo extends JpaRepository<Event, Long> {
 }
